Reorganize routes folder some more
Implement get feeds route

Implement put feeds

Implement delete feed route

Fix tests

Force color

More color tweaks

Add why is node running

Remove log

Fix tests?

Fix?

Fix route params

Wait for app

Remove unneeded ends
