+++ routes/api/bookmarks/_id/get-bookmark.js
@@ -1,5 +1,5 @@
 /* eslint-disable camelcase */
-import { getBookmarksQuery } from '../get-bookmark-query.js'
+import { getBookmarksQuery } from '../get-bookmarks-query.js'
 import { fullBookmarkProps } from '../bookmark-props.js'
 
 export async function getBookmark (fastify, opts) {
@@ -33,8 +33,6 @@ export async function getBookmark (fastify, opts) {
         bookmarkId,
         perPage: 1
       })
-
-      console.log(query)
 
       const results = await fastify.pg.query(query)
       const bookmark = results.rows[0]
