+++ routes/api/bookmarks/_id/put-bookmark.js
@@ -1,6 +1,6 @@
 /* eslint-disable camelcase */
 import SQL from '@nearform/sql'
-import { createEpisode } from '../../../../lib/create-episode.js'
+import { createEpisode } from '../../episodes/create-episode-query.js'
 import { runYTDLP } from '../../../../lib/run-yt-dlp.js'
 import { commnonBookmarkProps } from '../bookmark-props.js'
