+++ routes/api/bookmarks/put-bookmarks.js
@@ -4,7 +4,7 @@ import {
   commnonBookmarkProps,
   createEpisodeProp
 } from './bookmark-props.js'
-import { createEpisode } from '../../../lib/create-episode.js'
+import { createEpisode } from '../episodes/create-episode-query.js'
 import { runYTDLP } from '../../../lib/run-yt-dlp.js'
 
 export async function putBookmarks (fastify, opts) {
