+++ routes/api/episodes/index.js
@@ -1,0 +1,1 @@
+export default async function episodesRoutes (fastify, opts) {}
