+++ test/routes/user.test.js
@@ -2,8 +2,7 @@ import { test } from 'tap'
 import { build } from '../helper.js'
 
 test('default root route', async (t) => {
-  const app = build(t)
-
+  const app = await build(t)
   const res = await app.inject({
     url: '/api/user'
   })
